feat: switch from container-interop to PSR-11
This patch switches from using container-interop to PSR-11.

There were three container-interop **implementations**:

- `Exception\NotFoundException`
- `Symmetric\SymmetricPluginManager`
- `Symmetric\PaddingPluginManager`

These are all internal classes.
The exception is thrown by the two plugin manager implementations, and those are straight-up `ContainerInterface` implementations.

The classes that consume the plugin managers were checking for container-interop `ContainerInterface` implementations previously, but now check for PSR-11 versions, which is considered a type-widening operation, and thus backwards compatible.
If we assume that the plugin managers are considered internal details, then this would be backwards compatible; if not, this is a BC break.

Signed-off-by: Matthew Weier O'Phinney <matthew@weierophinney.net>